Fix rP22668 Wind computation - switch to signed integers.
Silences a GCC 9 warning and the logic was faulty.

Reported by: elexis

Differential Revision: https://code.wildfiregames.com/D2180


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@22670 3db68df2-c116-0410-a063-a993310a9797

+++ source/renderer/WaterManager.cpp
@@ -1045,8 +1045,8 @@ void WaterManager::RecomputeWindStrength()
 
 	struct SWindPoint {
 		SWindPoint(size_t x, size_t y, float strength) : X(x), Y(y), windStrength(strength) {}
-		size_t X;
-		size_t Y;
+		ssize_t X;
+		ssize_t Y;
 		float windStrength;
 	};
 
@@ -1128,7 +1128,7 @@ void WaterManager::RecomputeWindStrength()
 				point.X += movement[step].first;
 				point.Y += movement[step].second;
 
-				if (point.X < 0 || point.X >= m_MapSize || point.Y < 0 || point.Y >= m_MapSize)
+				if (point.X < 0 || point.X >= static_cast<ssize_t>(m_MapSize) || point.Y < 0 || point.Y >= static_cast<ssize_t>(m_MapSize))
 				{
 					onMap = false;
 					break;
